feat: move types and utils to external files
